Создан класс Player
