+++ Game/index.js
@@ -1,0 +1,1 @@
+class Game {}
