--- random.js
@@ -1,3 +1,0 @@
-export const getRandom = num => {
-	return Math.ceil(Math.random() * num)
-}
